# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Survey API
  version: 1.0.0
  description: API для сервиса опросов и голосований
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
    RegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        displayName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    CreateSurveyRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
    CreateQuestionRequest:
      type: object
      required: [prompt, type]
      properties:
        prompt:
          type: string
        type:
          type: string
          enum: [single_choice, multiple_choice, text]
        isRequired:
          type: boolean
        position:
          type: integer
    CreateOptionRequest:
      type: object
      required: [label]
      properties:
        label:
          type: string
        position:
          type: integer
    SubmitResponseRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
            required: [questionId]
            properties:
              questionId:
                type: string
              textValue:
                type: string
              optionIds:
                type: array
                items:
                  type: string
paths:
  /health:
    get:
      summary: Healthcheck
      responses:
        "200":
          description: OK
  /openapi.yaml:
    get:
      summary: OpenAPI specification
      responses:
        "200":
          description: YAML document
  /api/v1/auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/me:
    get:
      summary: Current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
  /api/v1/surveys:
    get:
      summary: List surveys
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: scope
          schema:
            type: string
            enum: [mine, active, completed]
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, published, closed]
        - in: query
          name: sort_by
          schema:
            type: string
            enum: [created_at, published_at, responses_count, title]
        - in: query
          name: sort_order
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Survey list
    post:
      summary: Create survey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurveyRequest"
      responses:
        "201":
          description: Draft survey created
  /api/v1/surveys/mine:
    get:
      summary: List current author surveys
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Survey list
  /api/v1/surveys/{surveyId}:
    get:
      summary: Get survey details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Survey details
    patch:
      summary: Update draft survey
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurveyRequest"
      responses:
        "200":
          description: Survey updated
    delete:
      summary: Delete draft survey
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Survey deleted
  /api/v1/surveys/{surveyId}/questions:
    post:
      summary: Add question to draft survey
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "201":
          description: Question created
  /api/v1/surveys/{surveyId}/questions/{questionId}:
    patch:
      summary: Update question in draft survey
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
        - in: path
          name: questionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Question updated
    delete:
      summary: Delete question from draft survey
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
        - in: path
          name: questionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Question deleted
  /api/v1/questions/{questionId}/options:
    post:
      summary: Add option to choice question
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: questionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOptionRequest"
      responses:
        "201":
          description: Option created
  /api/v1/questions/{questionId}/options/{optionId}:
    patch:
      summary: Update option
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: questionId
          required: true
          schema:
            type: string
        - in: path
          name: optionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Option updated
    delete:
      summary: Delete option
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: questionId
          required: true
          schema:
            type: string
        - in: path
          name: optionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Option deleted
  /api/v1/surveys/{surveyId}/publish:
    post:
      summary: Publish draft survey
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Survey published
  /api/v1/surveys/{surveyId}/close:
    post:
      summary: Close published survey
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Survey closed
  /api/v1/public/surveys/{surveyId}:
    get:
      summary: Get published survey for participation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Survey details
  /api/v1/public/surveys/{surveyId}/responses:
    post:
      summary: Submit survey response
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitResponseRequest"
      responses:
        "201":
          description: Response submitted
  /api/v1/surveys/{surveyId}/analytics:
    get:
      summary: Get survey analytics
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Aggregated analytics
  /api/v1/surveys/{surveyId}/analytics/export:
    get:
      summary: Export analytics and raw responses
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analytics export